+++ zopeext/sphinxcontrib/zopeext/autointerface.py
@@ -30,7 +30,9 @@ The :dir:`autointerface` directive has the same form and option as the
    documenter.
 """
 import sphinx.ext.autodoc
-import sphinx.directives
+import sphinx.domains.python
+import sphinx.roles
+from sphinx.locale import l_
 
 import zope.interface.interface
 
@@ -38,13 +40,11 @@ def interface_getattr(*v):
     """Behaves like `getattr` but for zope Interface objects which
     hide the attributes.
 
-
-    .. note::  Originally I simply tried to override
-      :meth:`InterfaceDocumenter.special_attrgetter` to deal with the
-      special access needs of :class:`Interface` objects, but found
-      that this is not intended to be overwritten.  Instead one should
-      register the special accessor  using
-      :func:`app.add_autodoc_attrgetter`.
+    .. note:: Originally I simply tried to override
+      :meth:`InterfaceDocumenter.special_attrgetter` to deal with the special
+      access needs of :class:`Interface` objects, but found that this is not
+      intended to be overwritten.  Instead one should register the special
+      accessor using :func:`app.add_autodoc_attrgetter`.
     """
     obj, name = v[:2]
     if "__dict__" == name:
@@ -54,7 +54,7 @@ def interface_getattr(*v):
     try:
         return getattr(obj, name)
     except AttributeError:
-        if name in obj.names():
+        if name in obj.names(all=True):
             return obj.get(name)
         elif 2 < len(v):
             return v[2]
@@ -83,7 +83,9 @@ class InterfaceDocumenter(sphinx.ext.autodoc.ClassDocumenter):
     """
     objtype = 'interface'               # Called 'autointerface'
 
-    priority = 10       # needs a higher priority than ClassDocumenter
+    # Since these have very specific tests, we give the classes defined here
+    # very high priority so that they override any other documenters.
+    priority = 100 + sphinx.ext.autodoc.ClassDocumenter.priority
 
     @classmethod
     def can_document_member(cls, member, membername, isattr, parent):
@@ -109,11 +111,21 @@ class InterfaceAttributeDocumenter(sphinx.ext.autodoc.AttributeDocumenter):
     interface attributes.
     """
     objtype = 'interfaceattribute'   # Called 'autointerfaceattribute'
-    directivetype = 'attribute'      # Formats as a 'class' for now
+    directivetype = 'attribute'      # Formats as a 'attribute' for now
+    priority = 100 + sphinx.ext.autodoc.AttributeDocumenter.priority
 
     @classmethod
     def can_document_member(cls, member, membername, isattr, parent):
-        return isinstance(member, zope.interface.interface.Attribute)
+        res = (isinstance(member, zope.interface.interface.Attribute)
+               and not isinstance(member, zope.interface.interface.Method))
+        return res
+
+    def add_content(self, more_content, no_docstring=False):
+        # Revert back to default since the docstring *is* the correct thing to
+        # display here.
+        sphinx.ext.autodoc.ClassLevelDocumenter.add_content(
+            self, more_content, no_docstring)
+
 
 class InterfaceMethodDocumenter(sphinx.ext.autodoc.MethodDocumenter):
     """
@@ -121,7 +133,8 @@ class InterfaceMethodDocumenter(sphinx.ext.autodoc.MethodDocumenter):
     interface attributes.
     """
     objtype = 'interfacemethod'   # Called 'autointerfacemethod'
-    directivetype = 'method'      # Formats as a 'class' for now
+    directivetype = 'method'      # Formats as a 'method' for now
+    priority = 100 + sphinx.ext.autodoc.MethodDocumenter.priority
 
     @classmethod
     def can_document_member(cls, member, membername, isattr, parent):
@@ -130,13 +143,14 @@ class InterfaceMethodDocumenter(sphinx.ext.autodoc.MethodDocumenter):
     def format_args(self):
         return interface_format_args(self.object)
 
-class InterfaceDirective(sphinx.directives.DescDirective):
-    """An 'interface' directive."""
+#class InterfaceDirective(sphinx.ext.autodoc.AutoDirective):
+class InterfaceDirective(sphinx.domains.python.PyClasslike):
+    r"""An `'interface'` directive."""
     def get_index_text(self, modname, name_cls):
-        if self.desctype == 'interface':
+        if self.objtype == 'interface':
             if not modname:
                 return '%s (built-in interface)' % name_cls[0]
-            return '%s (interface in %s)' % (name_cls[0], modname)
+            return '%s (%s interface)' % (name_cls[0], modname)
         else:
             return ''
 
@@ -146,4 +160,9 @@ def setup(app):
     app.add_autodocumenter(InterfaceDocumenter)
     app.add_autodocumenter(InterfaceAttributeDocumenter)
     app.add_autodocumenter(InterfaceMethodDocumenter)
-    app.add_directive('interface', InterfaceDirective)
+
+    domain = sphinx.domains.python.PythonDomain
+    domain.object_types['interface'] = sphinx.domains.python.ObjType(
+        l_('interface'), 'interface', 'obj')
+    domain.directives['interface'] = InterfaceDirective
+    domain.roles['interface'] = sphinx.domains.python.PyXRefRole()