+++ issuetracker/CHANGES.rst
@@ -1,3 +1,60 @@
+0.7.2 (Mar 10, 2011)
+====================
+
+- #13: Fixed source distribution to include tests again
+- Fixed extraction of issue state for open issues from bitbucket
+- Ignore references in inline literals and literal blocks
+
+
+0.7.1 (Jan 19, 2011)
+====================
+
+- Fixed #8: Copy the stylesheet after build again to avoid exceptions on
+  non-existing build directories
+
+
+0.7 (Jan 08, 2011)
+==================
+
+- Issue information is now cached
+- Custom issue trackers must now connect to the ``issuetracker-resolve-issue``
+  event, the builtin ``missing-reference`` event is no longer used.
+
+
+0.6 (Jan 04, 2011)
+==================
+
+- Added support for the debian bugtracker (thanks to Fladischer Michael)
+- Fixed NameError in launchpad issue tracker
+- Bitbucket also uses HTTPS urls now
+
+
+0.5.4 (Nov 15, 2010)
+====================
+
+- Github uses HTTPS urls now
+
+
+0.5.3 (Nov 14, 2010)
+====================
+
+- Added license text to source tarball
+
+
+0.5.2 (Sep 17, 2010)
+====================
+
+- Issue reference resolvers get the application object now as fourth
+  argument.  The environment is availabe in the ``.env`` attribute of this
+  object.
+- #4: Fixed the URL of Google Code issues (thanks to Denis Bilenko)
+- Fixed detection of closed issues in Google Code (thanks to Denis Bilenko)
+- Improved error message, if ``issuetracker_issue_pattern`` has too many
+  groups (thanks to Denis Bilenko)
+- Added warnings for unexpected HTTP status codes in BitBucket and Google
+  Code issue trackers
+
+
 0.5.1 (Jul 25, 2010)
 ====================